requirements.txt
fastapi>=0.108.0
uvicorn>=0.27.0
inngest==0.5.11
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.2.0
qdrant-client>=1.8.0
numpy>=1.26.4
pydantic>=2.11.0
llama-index>=0.6.13
